+++ yash-builtin/CHANGELOG.md
@@ -13,6 +13,13 @@ Terminology: A _public dependency_ is one that’s exposed through this crate’
 public API (e.g., re-exported types).
 A _private dependency_ is used internally and not visible to downstream users.
 
+## [0.21.0] - Unreleased
+
+### Changed
+
+- Public dependency versions:
+    - yash-semantics (optional) 0.18.0 → 0.19.0
+
 ## [0.20.0] - 2026-07-12
 
 ### Added
@@ -883,6 +890,7 @@ The `wait` built-in no longer treats suspended jobs as terminated jobs.
 
 - Initial implementation of the `yash-builtin` crate
 
+[0.21.0]: https://github.com/magicant/yash-rs/releases/tag/yash-builtin-0.21.0
 [0.20.0]: https://github.com/magicant/yash-rs/releases/tag/yash-builtin-0.20.0
 [0.19.0]: https://github.com/magicant/yash-rs/releases/tag/yash-builtin-0.19.0
 [0.18.2]: https://github.com/magicant/yash-rs/releases/tag/yash-builtin-0.18.2